--- lib/cache.js
@@ -1,1 +1,0 @@
-module.exports = new (require('fragment-cache'))();
