+++ lib/utils.js
@@ -1,16 +1,39 @@
 'use strict';
 
 var regex = require('regex-not');
+var Cache = require('fragment-cache');
+var utils = module.exports;
+var cache = utils.cache = new Cache();
 
 /**
  * Cast `val` to an array
  * @return {Array}
  */
 
-exports.arrayify = function(val) {
+utils.arrayify = function(val) {
   if (!Array.isArray(val)) {
     return [val];
   }
+  return val;
+};
+
+/**
+ * Memoize a generated regex or function
+ */
+
+utils.memoize = function(type, pattern, options, fn) {
+  var key = utils.createKey(type + pattern, options);
+
+  if (cache.has(type, key)) {
+    return cache.get(type, key);
+  }
+
+  var val = fn(pattern, options);
+  if (options && options.cache === false) {
+    return val;
+  }
+
+  cache.set(type, key, val);
   return val;
 };
 
@@ -20,7 +43,7 @@ exports.arrayify = function(val) {
  * to the pattern string.
  */
 
-exports.createKey = function(pattern, options) {
+utils.createKey = function(pattern, options) {
   var key = pattern;
   if (typeof options === 'undefined') {
     return key;
@@ -35,7 +58,7 @@ exports.createKey = function(pattern, options) {
  * Create the regex to use for matching text
  */
 
-exports.createRegex = function(str) {
+utils.createRegex = function(str) {
   var opts = {contains: true, strictClose: false};
   return regex(str, opts);
 };
